feat: Add functionality to generate monthly calendars
This commit adds functions to generate monthly calendars in addition to the existing functionality of generating weekly calendars. The "get_month" function retrieves a list of weeks within a given month, and the "generate_monthly_html" function generates the HTML content for a monthly calendar using a Jinja2 template. The new template file "templates/monthly.html" is also included in this commit.